+++ doc/changelog.md
@@ -4,6 +4,9 @@
 
 ## New features
 
+- The deck import wizard can now directly download Scryfall search queries as deck lists
+    - Added a text field to enter a Scryfall card search query, a button to show the result on the Scryfall website,
+      and a button that downloads the search result as a deck list.
 - Localization support. Translations are managed on [Crowdin.com](https://crowdin.com/project/mtgproxyprinter).   
   Join, if you want to contribute translations :)
     - Interface language is chosen based on the system locale, but can be explicitly set in the application settings.
@@ -18,6 +21,8 @@
 
 ## Changed features
 
+- The deck list import wizard now supports downloading links from the Scryfall API card search at 
+  [https://api.scryfall.com/cards/search](https://scryfall.com/docs/api/cards/search) 
 - Support decimal values in document settings, like margins, image spacings and the card bleed width.
 - As a safety measure against DoS-attacks via loading malicious documents, limit numerical document settings
   to 10000mm. Limiting the paper size to 10m (~394in) in each direction prevents the creation
@@ -26,6 +31,10 @@
 - Improved the related card search: The search now finds tokens created by Dungeons.
   Right-clicking a card with "Venture" or "Initiative" now also suggests the tokens created by the dungeon rooms.
     - When updating from previous versions, this change takes effect after the next card data update.
+
+## Fixed issues
+
+- Improved performance of the image cleanup wizard, when the user has many images of non-English cards stored on disk.
 - Reworded and clarified some texts in the user interface
 
 # Version 0.28.3 (2024-07-07)  <a name="v0_28_3"></a>